+++ examples/lib/src/files_json.dart
@@ -23,6 +23,7 @@ const List<String> filesJson = [
   "webgl_geometry_convex",
   'webgl_geometry_nurbs',
   'webgl_geometry_spline_editor',
+  'webgl_geometries_parametric',
   'webgl_interactive_points',
   'webgl_interactive_raycasting_points',
   'webgl_interactive_lines',