+++ packages/three_js_geometry/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.1
+
+* Added Parametric Geometries.
+
 ## 0.1.0
 
 * Updated to newest version of three_js core, math, and curves.